♻️ Integrate versioning and is_versioned
