+++ lamindb/models/ulabel.py
@@ -14,13 +14,10 @@ from lamindb.base.fields import (
 from lamindb.errors import FieldValidationError
 
 from ..base.ids import base62_8
-from .base import (
-    LinkORM,
-)
 from .can_curate import CanCurate
 from .feature import Feature
 from .has_parents import HasParents
-from .record import BasicRecord, Record, _get_record_kwargs
+from .record import BasicRecord, LinkORM, Record, _get_record_kwargs
 from .run import Run, TracksRun, TracksUpdates, User, current_user_id
 from .transform import Transform
